Numerical Analysis

This module provides functionality to perform numerical analysis on a JuMP model. This module follows the main API and is activated by the struct:

MathOptAnalyzer.Numerical.Analyzer — Type

Analyzer() <: MathOptAnalyzer.AbstractAnalyzer

The Analyzer type is used to analyze the coefficients of a model for numerical issues.

Example

julia> data = MathOptAnalyzer.analyze(
    MathOptAnalyzer.Numerical.Analyzer(),
    model;
    threshold_dense_fill_in = 0.10,
    threshold_dense_entries = 1000,
    threshold_small = 1e-5,
    threshold_large = 1e+5,
)

The additional parameters:

threshold_dense_fill_in: The threshold for the fraction of non-zero entries in a constraint to be considered dense.

threshold_dense_entries: The minimum number of non-zero entries for a constraint to be considered dense.

threshold_small: The threshold for small coefficients in the model.

threshold_large: The threshold for large coefficients in the model.

The analysis will return issues of the abstract type:

MathOptAnalyzer.Numerical.AbstractNumericalIssue — Type

AbstractNumericalIssue <: AbstractNumericalIssue

Abstract type for numerical issues found during the analysis of a model.

Specifically the possible issues are:

MathOptAnalyzer.Numerical.VariableNotInConstraints — Type

VariableNotInConstraints <: AbstractNumericalIssue

The VariableNotInConstraints issue is identified when a variable appears in no constraints.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.VariableNotInConstraints)

MathOptAnalyzer.Numerical.EmptyConstraint — Type

EmptyConstraint <: AbstractNumericalIssue

The EmptyConstraint issue is identified when a constraint has no coefficients different from zero.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.EmptyConstraint)

MathOptAnalyzer.Numerical.VariableBoundAsConstraint — Type

VariableBoundAsConstraint <: AbstractNumericalIssue

The VariableBoundAsConstraint issue is identified when a constraint is equivalent to a variable bound, that is, the constraint has only one non-zero coefficient, and this coefficient is equal to one.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.VariableBoundAsConstraint)

MathOptAnalyzer.Numerical.DenseConstraint — Type

DenseConstraint <: AbstractNumericalIssue

The DenseConstraint issue is identified when a constraint has a fraction of non-zero entries greater than threshold_dense_fill_in and the number of non-zero entries is greater than threshold_dense_entries.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.DenseConstraint)

MathOptAnalyzer.Numerical.SmallMatrixCoefficient — Type

SmallMatrixCoefficient <: AbstractNumericalIssue

The SmallMatrixCoefficient issue is identified when a matrix coefficient in a constraint is smaller than threshold_small.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.SmallMatrixCoefficient)

MathOptAnalyzer.Numerical.LargeMatrixCoefficient — Type

LargeMatrixCoefficient <: AbstractNumericalIssue

The LargeMatrixCoefficient issue is identified when a matrix coefficient in a constraint is larger than threshold_large.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.LargeMatrixCoefficient)

MathOptAnalyzer.Numerical.SmallBoundCoefficient — Type

SmallBoundCoefficient <: AbstractNumericalIssue

The SmallBoundCoefficient issue is identified when a variable's bound (coefficient) is smaller than threshold_small.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.SmallBoundCoefficient)

MathOptAnalyzer.Numerical.LargeBoundCoefficient — Type

LargeBoundCoefficient <: AbstractNumericalIssue

The LargeBoundCoefficient issue is identified when a variable's bound (coefficient) is larger than threshold_large.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.LargeBoundCoefficient)

MathOptAnalyzer.Numerical.SmallRHSCoefficient — Type

SmallRHSCoefficient <: AbstractNumericalIssue

The SmallRHSCoefficient issue is identified when the right-hand-side (RHS) coefficient of a constraint is smaller than threshold_small.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.SmallRHSCoefficient)

MathOptAnalyzer.Numerical.LargeRHSCoefficient — Type

LargeRHSCoefficient <: AbstractNumericalIssue

The LargeRHSCoefficient issue is identified when the right-hand-side (RHS) coefficient of a constraint is larger than threshold_large.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.LargeRHSCoefficient)

MathOptAnalyzer.Numerical.SmallObjectiveCoefficient — Type

SmallObjectiveCoefficient <: AbstractNumericalIssue

The SmallObjectiveCoefficient issue is identified when a coefficient in the objective function is smaller than threshold_small.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.SmallObjectiveCoefficient)

MathOptAnalyzer.Numerical.LargeObjectiveCoefficient — Type

LargeObjectiveCoefficient <: AbstractNumericalIssue

The LargeObjectiveCoefficient issue is identified when a coefficient in the objective function is larger than threshold_large.

For more information, run:

julia> MathOptAnalyzer.summarize(MathOptAnalyzer.Numerical.LargeObjectiveCoefficient)

MathOptAnalyzer.Numerical.SmallObjectiveQuadraticCoefficient — Type

SmallObjectiveQuadraticCoefficient <: AbstractNumericalIssue

The SmallObjectiveQuadraticCoefficient issue is identified when a quadratic coefficient in the objective function is smaller than threshold_small.

For more information, run:

julia> MathOptAnalyzer.summarize(
    MathOptAnalyzer.Numerical.SmallObjectiveQuadraticCoefficient
)

MathOptAnalyzer.Numerical.LargeObjectiveQuadraticCoefficient — Type

LargeObjectiveQuadraticCoefficient <: AbstractNumericalIssue

The LargeObjectiveQuadraticCoefficient issue is identified when a quadratic coefficient in the objective function is larger than threshold_large.

For more information, run:

julia> MathOptAnalyzer.summarize(
    MathOptAnalyzer.Numerical.LargeObjectiveQuadraticCoefficient
)

MathOptAnalyzer.Numerical.SmallMatrixQuadraticCoefficient — Type

SmallMatrixQuadraticCoefficient <: AbstractNumericalIssue

The SmallMatrixQuadraticCoefficient issue is identified when a quadratic coefficient in a constraint is smaller than threshold_small.

For more information, run:

julia> MathOptAnalyzer.summarize(
    MathOptAnalyzer.Numerical.SmallMatrixQuadraticCoefficient
)

MathOptAnalyzer.Numerical.LargeMatrixQuadraticCoefficient — Type

LargeMatrixQuadraticCoefficient <: AbstractNumericalIssue

The LargeMatrixQuadraticCoefficient issue is identified when a quadratic coefficient in a constraint is larger than threshold_large.

For more information, run:

julia> MathOptAnalyzer.summarize(
    MathOptAnalyzer.Numerical.LargeMatrixQuadraticCoefficient
)

MathOptAnalyzer.Numerical.NonconvexQuadraticObjective — Type

NonconvexQuadraticObjective <: AbstractNumericalIssue

The NonconvexQuadraticObjective issue is identified when a quadratic objective function is non-convex.

For more information, run:

julia> MathOptAnalyzer.summarize(
    MathOptAnalyzer.Numerical.NonconvexQuadraticObjective
)

MathOptAnalyzer.Numerical.NonconvexQuadraticConstraint — Type

NonconvexQuadraticConstraint

The NonconvexQuadraticConstraint issue is identified when a quadratic constraint is non-convex.

For more information, run:

julia> MathOptAnalyzer.summarize(
    MathOptAnalyzer.Numerical.NonconvexQuadraticConstraint
)

These issues are saved in the data structure that is returned from the MathOptAnalyzer.analyze function:

MathOptAnalyzer.Numerical.Data — Type

Data

The Data structure holds the results of the analysis performed by the MathOptAnalyzer.Numerical.Analyzer. It contains various thresholds and the information about the model's variables, constraints, and objective function.